Switched to floats instead of ints

+++ src/castro/cshop/items/ChatColors.java
@@ -13,13 +13,15 @@ public class ChatColors extends ShopItemPermission
 	@Override
     public float getPricePerHour()
     {
+		//return 100/day;
 		return 3000 / month;
     }
 	
 	
 	@Override
-	public int minTime()
+	public float minTime()
 	{
+		//return hour;
 		return 24*hour;
 	}
 }
