+++ src/castro/cshop/items/base/ShopItem.java
@@ -7,10 +7,10 @@ import castro.cshop.Plugin;
 public abstract class ShopItem
 {
 	protected final static Plugin plugin = Plugin.get;
-	protected final static int hour  = 1;
-	protected final static int day   = 24*hour;
-	protected final static int week  = 7*day;
-	protected final static int month = 30*day; 
+	protected final static float hour  = 1;
+	protected final static float day   = 24*hour;
+	protected final static float week  = 7*day;
+	protected final static float month = 30*day; 
 	
 	public abstract float getPricePerHour();
 	public abstract boolean giveItem(Player player, ShopItemData itemData);
@@ -22,7 +22,7 @@ public abstract class ShopItem
 		return false;
 	}
 	
-	public int minTime()
+	public float minTime()
 	{
 		return 1*hour;
 	}
